feat: Upload recipes to database
Major changes:
    - Created a form for everything needed to post a recipe with form validation
        + Optional image
        + Required title
        + Required description
        + Requires at least one ingredient
        + Requires at least one step in the directions section
        + Adding tags are optional
    - Every recipe has an image, but if the user doesn't choose one then the recipe will use a default.
    - Recipe images are uploaded to Cloudinary. The Cloudinary link is saved to the database.
    - All tags in the Tag table are unique. Many recipes can share the same tag.
    - All tags are also made lowercase and don't allow spaces to avoid "duplicates" in the tag table.
    - The entities Direction, Ingredient, and TaggedRecipe where given there own embeddable composite key class.
    - Users need to be logged in to create a post.
    - A success modal appears when the form is successfully submitted.

Other changes:
    - Updated the error messages on every form to use Formik's ErrorMessage component instead of the previous implementation.

Addresses: Issue #11

+++ springboot-backend/src/main/java/com/example/springbootbackend/auth/AccountDeleteRequest.java
@@ -4,32 +4,16 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class AccountDeleteRequest {
 
-//    @JsonProperty("data")
-//    AccountInfo data;
-
-    private int id;
+    private Long id;
     private String password;
 
     public AccountDeleteRequest() { }
 
-//    public AccountDeleteRequest(AccountInfo data) {
-//        this.data = data;
-//    }
-//
-//    public AccountInfo getData() {
-//        return data;
-//    }
-//
-//    public void setData(AccountInfo data) {
-//        this.data = data;
-//    }
-
-
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
