+++ springboot-backend/src/main/java/com/example/springbootbackend/auth/ChangePasswordRequest.java
@@ -3,13 +3,13 @@ package com.example.springbootbackend.auth;
 public class ChangePasswordRequest {
 
     private String verificationCode;
-    private int userId;
+    private Long userId;
     private String newPassword;
     private String oldPassword;
 
     public ChangePasswordRequest() { }
 
-    public ChangePasswordRequest(String verificationCode, int userId, String newPassword) {
+    public ChangePasswordRequest(String verificationCode, Long userId, String newPassword) {
         this.verificationCode = verificationCode;
         this.userId = userId;
         this.newPassword = newPassword;
@@ -23,11 +23,11 @@ public class ChangePasswordRequest {
         this.verificationCode = verificationCode;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
