+++ springboot-backend/src/main/java/com/example/springbootbackend/repository/EmailTokenRepository.java
@@ -1,6 +1,5 @@
 package com.example.springbootbackend.repository;
 
-import com.example.springbootbackend.model.Account;
 import com.example.springbootbackend.model.EmailVerificationToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -10,13 +9,13 @@ import org.springframework.data.repository.query.Param;
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 
-public interface EmailTokenRepository extends JpaRepository<EmailVerificationToken, Integer> {
+public interface EmailTokenRepository extends JpaRepository<EmailVerificationToken, Long> {
 
     @Query("SELECT e FROM EmailVerificationToken e WHERE e.account.email = ?1")
     EmailVerificationToken findByEmail(String email);
 
     @Query("SELECT e FROM EmailVerificationToken e WHERE e.account.id = ?1")
-    EmailVerificationToken findByAccountId(Integer accountId);
+    EmailVerificationToken findByAccountId(Long accountId);
 
     @Query("SELECT e FROM EmailVerificationToken e WHERE e.code = :code")
     EmailVerificationToken findByCode(@Param("code") String code);
